pyproject.toml
[project]
name = "helt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "geopy>=2.4.1",
    "google-generativeai>=0.3.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "streamlit>=1.30.0",
]
